Add JavaFX tutorial part 4
