+++ data/dino.txt
@@ -1,1 +1,3 @@
  T | 0 | sleep
+ T | 0 | read
+ E | 0 | lunch | from: Mar 04 2023 12:34 to: Apr 03 2034 17:00
